+++ src/contexts/edited_text.rs
@@ -10,7 +10,7 @@ pub struct EditedTextContext {
     /// A mock bot with all API methods.
     pub bot: Arc<MockBot>,
     /// Id of the message.
-    pub message_id: i32,
+    pub message_id: u32,
     /// The sender of the message.
     pub from: Option<types::User>,
     /// The time the message was sent at.
@@ -31,7 +31,7 @@ pub struct EditedTextContext {
 impl EditedTextContext {
     pub(crate) fn new(
         bot: Arc<MockBot>,
-        message_id: i32,
+        message_id: u32,
         from: Option<types::User>,
         date: i64,
         chat: types::raw::Chat,
@@ -63,8 +63,8 @@ impl<'a> traits::ChatMethods<'a> for EditedTextContext {
         self.chat.id
     }
 
-    fn message_id(&self) -> u64 {
-        self.message_id as u64
+    fn message_id(&self) -> u32 {
+        self.message_id
     }
 }
