+++ src/types/video.rs
@@ -8,15 +8,15 @@ pub struct Video {
     /// The file ID of the video.
     pub file_id: String,
     /// The width of the video.
-    pub width: i64,
+    pub width: u32,
     /// The height of the video.
-    pub height: i64,
+    pub height: u32,
     /// The duration of the video.
-    pub duration: i64,
+    pub duration: u32,
     /// The thumb of the video.
     pub thumb: Option<PhotoSize>,
     /// The MIME type of the video.
     pub mime_type: Option<String>,
     /// The video file's size.
-    pub file_size: Option<i64>,
+    pub file_size: Option<u32>,
 }
